requirements.txt
fastapi>=0.115.2
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.2
webdriver-manager>=4.0.1
openai>=1.81.0
python-multipart>=0.0.18
jinja2>=3.1.2
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.10.0
httpx>=0.27.0
websockets>=12.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Additional dependencies for enhanced functionality
aiohttp>=3.8.0
typing-extensions>=4.0.0

# Ollama and AI Agent dependencies
ollama>=0.1.7
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.3.7

# PDF Processing dependencies
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-magic>=0.4.27
